Reduce complexity in our env-template implementation.
Currently we see the following report:

> Line 133: warning: cyclomatic complexity 19 of function
> (*envTemplateCommand).expandFile() is high (> 15) (gocyclo)

That's largely caused by having an inline map of functions for
our template-helpers.

Move those into their own file, and add them by name, reduces
the complexity and simplifies the code.

Also added examples of some of the advanced functionality to the
sample template-file.