Better handling of errors for unknown tokens.
Our lexer has a simple job, which is to read the tokens from
the input stream and return objects to describe each distinct
entity.

We handle this in the simple way:

* Look for symbols (+, -, etc)
* Look for numbers.
* Look for whitespace.
* If none of those matched
  * Look for identifiers

That last step was a little broken for us, we'd loop forward
over the inputstream and peel of characters that were permitted
in identifiers - if we found one that wasn't we'd break, and not
advance.

The net result was that input such as "#" or "10!" would result
in an infinite loop.  Each time trying to read "#" or "!" as
an identifier which wasn't a permitted identifier character.

This commit explicitly looks for invalid-characters and returns
an appropriate error.