BUGFIX: json-validator was using the wrong-glob
Due to copy/paste the JSON validator was attempting to walk the
filesystem looking for files matching the globs `*.yaml` &
`*.yml` , rather than `*.json` is it should have done.

If you specified explicitly the files you wished to process all
was OK, but without this fix using the validator would process
the wrong files.