Added new 'todo' sub-command.
This command is designed to output (comments) TODO/FIXME notes from
files which contain an associated data in the past.

The idea is that you can write things such as:

        // TODO (03/2022) - This must be fixed before March 2022

Now run `sysbox todo` and you'll see:

        $ sysbox todo
        ...
        foo.go: //TODO (03/2022) - This must be fixed before March 2022

We allow reporting on lines that contain either `FIXME` or `TODO`,
along with a corresponding date entry.  A date may be written as:

* (DD/MM/YYYY)
* (MM/YYYY)
* (YYYY)

Note: in all cases the "(" and ")" characters must be present.

This is inspired by an issue I read on a different project:

* https://github.com/cretz/software-ideas/issues/111